Move addCommas to utils
